Receive IP packets
We can't use SOCK_RAW and UDP or TCP, so we have to use the experimental
protocol numbers instead. This adds support for opening a SOCK_RAW
socket with protocol 253 or 254, as well as just enough IP packet
parsing support to read the protocol number and other fields necessary
to higher level checksum algorithms.